Module varisat::checker

Check unsatisfiability proofs.

Modules

write_lrat
Proof processor that generates an LRAT proof.

Structs

Checker
A checker for unsatisfiability proofs in the native varisat format.

Clause
Literals and metadata for non-unit clauses.

ClauseLits
Literals of a clause, either inline or an index into a buffer

TraceItem
Propagation of the RUP check.

UnitClause
Known unit clauses and metadata.

WriteLrat
Proof processor that generates an LRAT proof.

Enums

CheckedProofStep
A single step of a proof.

CheckerError
Possible errors while checking a varisat proof.

DeleteClauseResult
Return type of Checker::delete_clause

StoreClauseResult
Return type of Checker::store_clause

UnitId
Identifies the origin of a unit clause.

Constants

INLINE_LITS

_DERIVE_failure_Fail_FOR_CheckerError

_DERIVE_failure_core_fmt_Display_FOR_CheckerError

Traits

ProofProcessor
Implement to process proof steps.

Functions

copy_canonical
Sort literals, remove duplicates and check for tautologic clauses.

is_subset
Test whether a set of literals is a (strict) subset of another set of literals