Struct varisat::checker::Checker

pub struct Checker<'a> {
    step: u64,
    next_clause_id: u64,
    literal_buffer: Vec<Lit>,
    garbage_size: usize,
    clauses: HashMap<u64, SmallVec<[Clause; 1]>>,
    unit_clauses: Vec<Option<UnitClause>>,
    trail: Vec<(Lit, Option<UnitClause>)>,
    unsat: bool,
    ended: bool,
    trace: Vec<TraceItem>,
    trace_edges: Vec<LitIdx>,
    trace_ids: Vec<u64>,
    processors: Vec<&'a mut dyn ProofProcessor>,
    unit_conflict: Option<[u64; 2]>,
    tmp: Vec<Lit>,
    hash_bits: u32,
    previous_irred_clause_id: Option<u64>,
    previous_irred_clause_lits: Vec<Lit>,
    assumptions: Vec<Lit>,
}

A checker for unsatisfiability proofs in the native varisat format.

Fields
step: u64

Current step number.

next_clause_id: u64

Next clause id to use.

literal_buffer: Vec<Lit>

Literal storage for clauses,

garbage_size: usize

Number of literals in the buffer which are from deleted clauses.

clauses: HashMap<u64, SmallVec<[Clause; 1]>>

Stores all known non-unit clauses indexed by their hash.

unit_clauses: Vec<Option<UnitClause>>

Stores known unit clauses and propagations during a clause check.

trail: Vec<(Lit, Option<UnitClause>)>

Stores overwritten values in unit_clauses to undo assignments.

unsat: bool

Whether unsatisfiability was proven.

ended: bool

Whether an end of proof step was checked.

trace: Vec<TraceItem>

Involved clauses during the last check.

trace_edges: Vec<LitIdx>

Edges of the trace implication graph.

trace_ids: Vec<u64>

Just the ids of trace.

processors: Vec<&'a mut dyn ProofProcessor>

Registered proof processors.

unit_conflict: Option<[u64; 2]>

This stores a conflict of input unit clauses.

Our representation for unit clauses doesn't support conflicting units so this is used as a workaround.

tmp: Vec<Lit>

Temporary storage for literals.

hash_bits: u32

How many bits are used for storing clause hashes.

previous_irred_clause_id: Option<u64>

Last added irredundant clause id.

Sorted and free of duplicates.

previous_irred_clause_lits: Vec<Lit>

Last added irredundant clause literals.

assumptions: Vec<Lit>

Current assumptions, used to check FailedAssumptions and Model

Methods

impl<'a> Checker<'a>

pub fn new() -> Checker<'a>

Create a new checker.

pub fn add_formula(&mut self, formula: &CnfFormula) -> Result<(), CheckerError>

Add a formula to the checker.

pub fn add_clause(&mut self, clause: &[Lit]) -> Result<(), CheckerError>

Adds a clause to the checker.

pub fn add_dimacs_cnf(&mut self, input: impl Read) -> Result<(), Error>

Reads and adds a formula in DIMACS CNF format.

Using this avoids creating a temporary CnfFormula.

fn clause_hash(&self, lits: &[Lit]) -> u64

Compute a clause hash of the current bit size

fn lit_value(&self, lit: Lit) -> Option<(bool, UnitClause)>

Value of a literal if known from unit clauses.

fn store_clause(
&mut self,
lits: &[Lit],
redundant: bool
) -> (u64, StoreClauseResult)

Adds a clause to the checker data structures.

lits must be sorted and free of duplicates.

Returns the id of the added clause and indicates whether the clause is new or changed from redundant to irredundant.

fn store_unit_clause(&mut self, lit: Lit) -> (u64, StoreClauseResult)

Adds a unit clause to the checker data structures.

Returns the id of the added clause and a boolean that is true if the clause wasn't already present.

fn delete_clause(
&mut self,
lits: &[Lit],
redundant: bool
) -> Result<(u64, DeleteClauseResult), CheckerError>

Delete a clause from the current formula.

lits must be sorted and free of duplicates.

Returns the id of the deleted clause and whether the ref_count (or irredundant ref_count) became zero.

fn collect_garbage(&mut self)

Perform a garbage collection if required

fn rehash(&mut self)

Recompute all clause hashes

fn check_clause_with_hashes(
&mut self,
lits: &[Lit],
propagation_hashes: &[u64]
) -> Result<(), CheckerError>

Check whether a clause is implied by clauses of the given hashes.

lits must be sorted and free of duplicates.

fn subsumed_by_previous_irred_clause(&self, lits: &[Lit]) -> bool

Check whether a given clause is subsumed by the last added irredundant clause.

lits must be sorted and free of duplicates.

pub(crate) fn check_step(&mut self, step: ProofStep) -> Result<(), CheckerError>

Check a single proof step

fn check_at_clause_step(
&mut self,
redundant: bool,
clause: &[Lit],
propagation_hashes: &[u64]
) -> Result<(), CheckerError>

Check an AtClause step

fn check_delete_clause_step(
&mut self,
clause: &[Lit],
proof: DeleteClauseProof
) -> Result<(), CheckerError>

Check a DeleteClause step

fn check_unit_clauses_step(
&mut self,
units: &[(Lit, u64)]
) -> Result<(), CheckerError>

Check a UnitClauses step

fn check_model_step(&mut self, model: &[Lit]) -> Result<(), CheckerError>

Check a Model step

fn check_failed_assumptions_step(
&mut self,
failed_core: &[Lit],
propagation_hashes: &[u64]
) -> Result<(), CheckerError>

Check a FailedAssumptions step

fn process_step<'b>(
processors: &'b mut [&'a mut dyn ProofProcessor],
step: &CheckedProofStep<'b>
) -> Result<(), CheckerError>

Invoke all proof processors for a CheckedProofStep

pub fn check_proof(&mut self, input: impl Read) -> Result<(), CheckerError>

Reads and adds a formula in DIMACS CNF format.

Using this avoids creating a temporary CnfFormula.

pub(crate) fn process_unit_conflicts(&mut self) -> Result<(), CheckerError>

Process unit conflicts detected during clause loading.

pub fn add_processor(&mut self, processor: &'a mut dyn ProofProcessor)

Add a ProofProcessor.

This has to be called before loading any clauses or checking any proofs.