Struct varisat::clause::Clause

#[repr(transparent)]

pub struct Clause {
    data: [LitIdx],
}

A clause.

This is stoed in a ClauseAlloc and thus must have a representation compatible with slice of LitIdx values.

It would be nicer to use a DST struct with two members and repr(C), but while that can be declared in stable rust, it's almost impossible to work with.

Fields
data: [LitIdx]
Methods

impl Clause

pub fn header(&self) -> &ClauseHeader

The clause's header

pub fn header_mut(&mut self) -> &mut ClauseHeader

Mutable reference to the clause's header

pub fn lits(&self) -> &[Lit]

The clause's literals

pub fn lits_mut(&mut self) -> &mut [Lit]

Mutable slice of the clause's literals