Struct varisat::context::Context

pub struct Context<'a> {
    pub analyze_conflict: AnalyzeConflict,
    pub assignment: Assignment,
    pub binary_clauses: BinaryClauses,
    pub clause_activity: ClauseActivity,
    pub clause_alloc: ClauseAlloc,
    pub clause_db: ClauseDb,
    pub impl_graph: ImplGraph,
    pub assumptions: Assumptions,
    pub model: Model,
    pub proof: Proof<'a>,
    pub schedule: Schedule,
    pub solver_config: SolverConfig,
    pub solver_state: SolverState,
    pub tmp_data: TmpData,
    pub tmp_flags: TmpFlags,
    pub trail: Trail,
    pub variables: Variables,
    pub vsids: Vsids,
    pub watchlists: Watchlists,
}

Central solver data structure.

This struct contains all data kept by the solver. Most functions operating on multiple fields of the context use partial references provided by the partial_ref crate. This documents the data dependencies and makes the borrow checker happy without the overhead of passing individual references.

Fields
analyze_conflict: AnalyzeConflictassignment: Assignmentbinary_clauses: BinaryClausesclause_activity: ClauseActivityclause_alloc: ClauseAllocclause_db: ClauseDbimpl_graph: ImplGraphassumptions: Assumptionsmodel: Modelproof: Proof<'a>schedule: Schedulesolver_config: SolverConfigsolver_state: SolverStatetmp_data: TmpDatatmp_flags: TmpFlagstrail: Trailvariables: Variablesvsids: Vsidswatchlists: Watchlists