Struct varisat::dimacs::DimacsParser

pub struct DimacsParser {
    formula: CnfFormula,
    partial_clause: Vec<Lit>,
    header: Option<DimacsHeader>,
    line_number: usize,
    clause_count: usize,
    partial_lit: usize,
    negate_next_lit: bool,
    in_lit: bool,
    in_comment_or_header: bool,
    in_header: bool,
    start_of_line: bool,
    error: bool,
    header_line: Vec<u8>,
}

Parser for DIMACS CNF files.

This parser can consume the input in chunks while also producing the parsed result in chunks.

Fields
formula: CnfFormulapartial_clause: Vec<Lit>header: Option<DimacsHeader>line_number: usizeclause_count: usizepartial_lit: usizenegate_next_lit: boolin_lit: boolin_comment_or_header: boolin_header: boolstart_of_line: boolerror: boolheader_line: Vec<u8>
Methods

impl DimacsParser

pub fn new() -> DimacsParser

Create a new DIMACS CNF parser.

pub fn parse<impl io::Read>(input: impl io::Read) -> Result<CnfFormula, Error> where
impl io::Read: Read,

Parse the given input and check the header if present.

This parses the whole input into a single CnfFormula. Incremental parsing is possible using parse_incremental or the parse_chunk method.

pub fn parse_incremental<impl io::Read, impl FnMut(&mut DimacsParser) -> Result<(), Error>>(
input: impl io::Read,
callback: impl FnMut(&mut DimacsParser) -> Result<(), Error>
) -> Result<DimacsParser, Error> where
impl FnMut(&mut DimacsParser) -> Result<(), Error>: FnMut(&mut DimacsParser) -> Result<(), Error>,
impl io::Read: Read,

Parse the given input incrementally and check the header if present.

The callback is invoked repeatedly with a reference to the parser. The callback can process the formula incrementally by calling take_formula on the passed argument.

pub fn parse_chunk(&mut self, chunk: &[u8]) -> Result<(), ParserError>

Parse a chunk of input.

After parsing the last chunk call the eof method.

If this method returns an error, the parser is in an invalid state and cannot parse further chunks.

pub fn eof(&mut self) -> Result<(), ParserError>

Finish parsing the input.

This does not check whether the header information was correct, call check_header for this.

pub fn check_header(&self) -> Result<(), ParserError>

Verifies the header information when present.

Does nothing when the input doesn't contain a header.

pub fn take_formula(&mut self) -> CnfFormula

Returns the subformula of everything parsed since the last call to this method.

To parse the whole input into a single CnfFormula, simply call this method once after calling eof. For incremental parsing this method can be invoked after each call of parse_chunk.

The variable count of the returned formula will be the maximum of the variable count so far and the variable count of the header if present.

pub fn header(&self) -> Option<DimacsHeader>

Return the DIMACS CNF header data if present.

pub fn clause_count(&self) -> usize

Number of clauses parsed.

pub fn var_count(&self) -> usize

Number of variables in the parsed formula.