Struct varisat::lit::Lit

#[repr(transparent)]

pub struct Lit {
    code: u32,
}

A boolean literal.

A literal is a variable or the negation of a variable.

Conceptually a literal consists of a Var and a bool indicating whether the literal represents the variable (positive literal) or its negation (negative literal).

Internally a literal is represented as an integer that is two times the index of its variable when it is positive or one more when it is negative. This integer is called the code of the literal.

The restriction on the range of allowed indices for Var also applies to Lit.

Fields
code: u32
Methods

impl Lit

pub fn from_var(var: Var, polarity: bool) -> Lit

Creates a literal from a Var and a bool that is true when the literal is positive.

pub fn positive(var: Var) -> Lit

Create a positive literal from a Var.

pub fn negative(var: Var) -> Lit

Create a negative literal from a Var.

pub fn from_index(index: usize, polarity: bool) -> Lit

Create a literal from a variable index and a bool that is true when the literal is positive.

pub fn from_code(code: usize) -> Lit

Create a literal with the given encoding.

pub fn from_dimacs(number: isize) -> Lit

Creates a literal from an integer.

The absolute value is used as 1-based index, the sign of the integer is used as sign of the literal.

pub fn to_dimacs(self) -> isize

1-based Integer representation of the literal, opposite of from_dimacs.

pub fn index(self) -> usize

0-based index of the literal's variable.

pub fn var(self) -> Var

The literal's variable.

pub fn is_negative(self) -> bool

Whether the literal is negative, i.e. a negated variable.

pub fn is_positive(self) -> bool

Whether the literal is positive, i.e. a non-negated variable.

pub fn code(self) -> usize

Two times the variable's index for positive literals and one more for negative literals.

This is also the internal encoding.

pub fn map_var<impl FnOnce(Var) -> Var>(self, f: impl FnOnce(Var) -> Var) -> Lit where
impl FnOnce(Var) -> Var: FnOnce(Var) -> Var,

Apply a function to the variable of the literal, without changing the polarity.