Module varisat::variables

Variable mapping and metadata.

Modules

data
Data associated with variables.

var_map
Mappings between variable names

Structs

Variables
Variable mapping and metadata.

Functions

delete_global_if_unused
Delete a global variable if it is unused

global_from_user
Maps a user variable into a global variable.

initialize_solver_var
Initialize a newly allocated solver variable

new_user_var
Allocates a currently unused user variable.

observe_internal_vars
Turns all hidden vars into witness vars and returns them.

remove_solver_var
Remove a solver var.

set_sampling_mode
Changes the sampling mode of a global variable.

solver_from_global
Maps an existing global variable to a solver variable.

solver_from_user
Maps a user variable to a solver variable.

solver_from_user_lits
Maps a slice of user lits to solver lits using solver_from_user.